refactor: extract StatsCard component
