+++ libs/defi/oeth/src/portfolio/views/PortfolioView.tsx
@@ -3,6 +3,7 @@ import {
   Page,
   PageSection,
   PageTitle,
+  StatsCard,
   trackSentryError,
 } from '@origin/defi/shared';
 import { ErrorBoundary, ErrorCard } from '@origin/shared/components';
@@ -10,7 +11,7 @@ import { tokens } from '@origin/shared/contracts';
 import { useIntl } from 'react-intl';
 
 import { OethHistoryCard } from '../components/OethHistoryCard';
-import { OethStats, WoethStats } from '../components/StatsCard';
+import { WoethStats } from '../components/StatsCard';
 import { TokenSwitch } from '../components/TokenSwitch';
 import { WoethHistoryCard } from '../components/WoethHistoryCard';
 import { useTokenSelect } from '../hooks';
@@ -35,7 +36,11 @@ export const PortfolioView = () => {
             ErrorComponent={<ErrorCard />}
             onError={trackSentryError}
           >
-            {token === 'oeth' ? <OethStats /> : <WoethStats />}
+            {token === 'oeth' ? (
+              <StatsCard token={tokens.mainnet.OETH} />
+            ) : (
+              <WoethStats />
+            )}
             {token === 'oeth' ? <OethHistoryCard /> : <WoethHistoryCard />}
           </ErrorBoundary>
         </Stack>
