+++ libs/defi/shared/src/components/Cards/index.ts
@@ -1,2 +1,3 @@
 export * from './BridgePromoCard';
 export * from './GlobalStatsCard';
+export * from './StatsCard';
